+++ src/onegov.town/onegov/town/models/__init__.py
@@ -5,6 +5,7 @@ from onegov.town.models.form import BuiltinFormDefinition, CustomFormDefinition
 from onegov.town.models.image import Image, ImageCollection, Thumbnail
 from onegov.town.models.page import News, Topic
 from onegov.town.models.resource import DaypassResource
+from onegov.town.models.sitecollection import SiteCollection
 from onegov.town.models.town import Town
 from onegov.town.models.traitinfo import TraitInfo
 
@@ -20,6 +21,7 @@ __all__ = [
     'Image',
     'ImageCollection',
     'News',
+    'SiteCollection',
     'Thumbnail',
     'Topic',
     'Town',